Refactor the authorization functions into a sub module, using flask blueprint
