+++ app/__init__.py
@@ -21,7 +21,7 @@ app.config.from_object(Config)
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 login = LoginManager(app)
-login.login_view = 'login'
+login.login_view = 'auth.login'
 mail = Mail(app)
 bootstrap = Bootstrap(app)
 moment = Moment(app)
@@ -31,6 +31,8 @@ babel = Babel(app)
 # Register blueprints
 from app.errors import bp as bp_errors
 app.register_blueprint(bp_errors)
+from app.auth import bp as bp_auth
+app.register_blueprint(bp_auth, url_prefix='/auth')
 
 
 from app import routes, models
