+++ app/templates/auth/login.html
@@ -11,12 +11,12 @@
     <br>
     <p>
         Forgot your password?
-        <a href="{{ url_for('request_password_reset') }}">
+        <a href="{{ url_for('auth.request_password_reset') }}">
             Click here to reset!
         </a>
     </p>
     <p>
         Not registered yet?
-        <a href="{{ url_for('register') }}">Click here to register</a>!
+        <a href="{{ url_for('auth.register') }}">Click here to register</a>!
     </p>
 {% endblock %}
