+++ app/templates/base.html
@@ -42,7 +42,7 @@
                 <ul class="nav navbar-nav navbar-right">
                     {% if current_user.is_anonymous %}
                     <li>
-                        <a href="{{ url_for('login') }}">
+                        <a href="{{ url_for('auth.login') }}">
                            {{ _('Login') }}
                         </a>
                     </li>
@@ -54,7 +54,7 @@
                         </a>
                     </li>
                     <li>
-                        <a href="{{ url_for('logout') }}">
+                        <a href="{{ url_for('auth.logout') }}">
                             {{ _('Logout') }}
                         </a>
                     </li>
